Flytt søknad progress-endepunkt til SøknadApi
